add context analyzers
