+++ lib/decidim/decidim_awesome.rb
@@ -4,14 +4,16 @@ require "decidim/decidim_awesome/railtie" if defined? Rails
 require "decidim/decidim_awesome/admin"
 require "decidim/decidim_awesome/engine"
 require "decidim/decidim_awesome/admin_engine"
+require "decidim/decidim_awesome/content_renderers"
+require "decidim/decidim_awesome/context_analyzers"
 
 module Decidim
   module DecidimAwesome
     include ActiveSupport::Configurable
 
     autoload :Config, "decidim/decidim_awesome/config"
-    autoload :ContextAnalyzer, "decidim/decidim_awesome/context_analyzer"
-    autoload :MarkdownRenderer, "decidim/decidim_awesome/content_renderers/markdown_renderer"
+    autoload :ContentRenderes, "decidim/decidim_awesome/content_renderers"
+    autoload :ContextAnalyzers, "decidim/decidim_awesome/context_analyzers"
 
     config_accessor :allow_images_in_full_editor do
       false
