+++ lib/decidim/decidim_awesome/content_renderers.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Decidim
+  module DecidimAwesome
+    module ContentRenderers
+      autoload :MarkdownRenderer, "decidim/decidim_awesome/content_renderers/markdown_renderer"
+    end
+  end
+end
