+++ lib/decidim/decidim_awesome/test/factories.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require "decidim/core/test/factories"
+require "decidim/proposals/test/factories"
 
 FactoryBot.define do
   factory :awesome_config, class: "Decidim::DecidimAwesome::AwesomeConfig" do
